% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch13-fn.R
\name{anova1.diff}
\alias{anova1.diff}
\title{CI for the difference in means: one-way ANOVA}
\usage{
anova1.diff(y, f, xl, yl, alp = 0.05, plot = TRUE, dig = 4)
}
\arguments{
\item{y}{Response variable data.}

\item{f}{Factors (same length as y).}

\item{xl}{Label of x-axis, Default: 'Factor'.}

\item{yl}{Label of y-axis, Default: 'Response Variable'.}

\item{alp}{Level of significance, Default: 0.05.}

\item{plot}{Logical value for plotting the confidence intervals (default=TRUE).}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
None.
}
\description{
Estimating the CI for difference between means in one-way ANOVA.
}
\examples{
y = c(79,83,88,78,75, 81,89,91,84,86,82, 86,91,93,90,89, 76,81,82,79)
f = c(rep(150,5), rep(200,6), rep(250,5), rep(300,4))
anova1.diff(y, f, xl="Temp", yl="Yield")

aa=anova1.fit(y, f, xl="Temp", yl="Yield")
anova1.diff(aa)
Temp=as.factor(f)
dd=aov(y ~ Temp)
anova1.diff(dd, yl="Yield")
}
