% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch13-fn.R
\name{anova2.conf}
\alias{anova2.conf}
\alias{anova2.diff}
\title{Estimating the Means: Two-Way ANOVA}
\usage{
anova2.conf(y, f1, f2, xl1, xl2, yl, alp = 0.05, plot = TRUE,
  inter = TRUE, maxim = TRUE, dig = 4)

anova2.diff(y, f1, f2, xl1, xl2, yl, alp = 0.05, plot = TRUE,
  inter = TRUE, maxim = TRUE, nb = 4, dig = 4)
}
\arguments{
\item{y}{Response variable data.}

\item{f1}{Levels of factor1 (same length as y).}

\item{f2}{Levels of factor2 (same length as y).}

\item{xl1}{Label of x-axis (default=name of f1).}

\item{xl2}{Legend label (default=name of f2).}

\item{yl}{Label of y-axis (default=name of y).}

\item{alp}{Level of significance, Default: 0.05.}

\item{plot}{Logical value for drawing diagnostic plots, Default: TRUE.}

\item{inter}{Logical value for including interaction, Default: TRUE.}

\item{maxim}{Logical value for maximization problem, Default: TRUE.}

\item{dig}{Number of digits below the decimal point, Default: 4.}

\item{nb}{Number of best level to be compared, Default: 4.}

\item{y}{Response variable data.}

\item{f1}{Levels of factor1 (same length as y).}

\item{f2}{Levels of factor2 (same length as y).}

\item{xl1}{Label of x-axis (default=name of f1).}

\item{xl2}{Legend label (default=name of f2).}

\item{yl}{Label of y-axis (default=name of y).}

\item{alp}{Level of significance, Default: 0.05.}

\item{plot}{Logical value for drawing diagnostic plots, Default: TRUE.}

\item{inter}{Logical value for including interaction, Default: TRUE.}

\item{maxim}{Logical value for maximization problem, Default: TRUE.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
None.

None.
}
\description{
Estimating the CI for the Means: Two-Way ANOVA.

Estimating the CI for the Difference in Means: Two-Way ANOVA.
}
\examples{
Yield = c(76,79,81,79,83,85, 79,81,84,86,89,88, 87,91,91,94,88,86, 79,82,85,84,77,76)
f1 = c(rep(100,6), rep(150,6), rep(200,6), rep(250,6))
f2 = rep(c(1,1,2,2,3,3), 4)
Temp = as.factor(f1)
Press = as.factor(paste0(f2, "Psig"))
anova2.conf(Yield, Temp, Press)

bb=anova2.fit(Yield, Temp, Press)
anova2.conf(bb)

ee=aov(Yield ~ Temp * Press)
anova2.conf(ee)

# Without replication
Yield1=c(77.5,80.0,89.0,80.5, 80.0,85.0,92.5,84.5, 84.0,88.5,87.0,76.5)
Temp1=c(100,150,200,250, 100,150,200,250, 100,150,200,250)
Press1=c(1,1,1,1, 2,2,2,2 ,3,3,3,3)
anova2.conf(Yield1, Temp1, Press1)

cc=anova2.fit(Yield1, Temp1, Press1)
anova2.conf(cc)

Temp.f=as.factor(Temp1)
Press.f=as.factor(Press1)
ff=aov(Yield1 ~ Temp.f + Press.f)
anova2.conf(ff)
Yield = c(76,79,81,79,83,85, 79,81,84,86,89,88, 87,91,91,94,88,86, 79,82,85,84,77,76)
f1 = c(rep(100,6), rep(150,6), rep(200,6), rep(250,6))
f2 = rep(c(1,1,2,2,3,3), 4)
Temp = as.factor(f1)
Press = as.factor(paste0(f2, "Psig"))
anova2.diff(Yield, Temp, Press)

bb=anova2.fit(Yield, Temp, Press)
anova2.diff(bb)

ee=aov(Yield ~ Temp * Press)
anova2.diff(ee)

# Without replication
Yield1=c(77.5,80.0,89.0,80.5, 80.0,85.0,92.5,84.5, 84.0,88.5,87.0,76.5)
Temp1=c(100,150,200,250, 100,150,200,250, 100,150,200,250)
Press1=c(1,1,1,1, 2,2,2,2 ,3,3,3,3)
anova2.diff(Yield1, Temp1, Press1)

cc=anova2.fit(Yield1, Temp1, Press1)
anova2.diff(cc)

Temp.f=as.factor(Temp1)
Press.f=as.factor(Press1)
ff=aov(Yield1 ~ Temp.f + Press.f)
anova2.diff(ff)
}
