% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch13-fn.R
\name{anova2.fit}
\alias{anova2.fit}
\title{Two-Way Analysis of Variance}
\usage{
anova2.fit(y, f1, f2, xl1, xl2, yl, alp = 0.05, iplot = TRUE, plot = TRUE,
  inter = TRUE, dig = 4)
}
\arguments{
\item{y}{Response variable data.}

\item{f1}{Levels of factor1 (same length as y).}

\item{f2}{Levels of factor2 (same length as y).}

\item{xl1}{Label of x-axis (default=name of f1).}

\item{xl2}{Legend label (default=name of f2).}

\item{yl}{Label of y-axis (default=name of y).}

\item{alp}{Level of significance, Default: 0.05.}

\item{iplot}{Logical value for drawing the interaction plot, Default: TRUE.}

\item{plot}{Logical value for drawing diagnostic plots, Default: TRUE.}

\item{inter}{Logical value for including interaction, Default: TRUE.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
None.
}
\description{
Two-way analysis of variance to produce an ANOVA table.
}
\examples{
Yield = c(76,79,81,79,83,85, 79,81,84,86,89,88, 87,91,91,94,88,86, 79,82,85,84,77,76)
f1 = c(rep(100,6), rep(150,6), rep(200,6), rep(250,6))
f2 = rep(c(1,1,2,2,3,3), 4)
Temp = as.factor(f1)
Press = as.factor(paste0(f2, "Psig"))
anova2.fit(Yield, Temp, Press)
# Without replication
Yield1=c(77.5,80.0,89.0,80.5, 80.0,85.0,92.5,84.5, 84.0,88.5,87.0,76.5)
Temp1=c(100,150,200,250, 100,150,200,250, 100,150,200,250)
Press1=c(1,1,1,1, 2,2,2,2 ,3,3,3,3)
anova2.fit(Yield1, Temp1, Press1)
}
