% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{chi.gof1}
\alias{chi.gof1}
\title{Goodness of Fit Test from a Table}
\usage{
chi.gof1(x, p, alp = 0.05, plot = TRUE, dig = 2, dig2 = 4)
}
\arguments{
\item{x}{Table of frequency}

\item{p}{Probability vector}

\item{alp}{Level of significance, Default: 0.05}

\item{plot}{Logical value for plotting the test result, Default: TRUE}

\item{dig}{Number of digits below the decimal point for the table, Default: 2}

\item{dig2}{Number of digits below the decimal point for the results, Default: 4}
}
\value{
list(stat=test stat, df=degree of freedom, crv=critical value, pv=p-value, tab=table)
}
\description{
Goodness of fit test from a table, when the probability is given.
}
\examples{
# Goodness-of-fit Test
x = c(31,26,22,18,13,10)
chi.gof1(x)

x = c(32, 65, 47, 38, 18)
p = c(0.15, 0.3, 0.25, 0.2, 0.1)
chi.gof1(x, p)
}
