% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{cross.test}
\alias{cross.test}
\title{Test of a Cross Table}
\usage{
cross.test(x, v1, v2, alp = 0.05, prt = TRUE, plot = TRUE, dig = 2,
  dig2 = 4)
}
\arguments{
\item{x}{Cross Table}

\item{v1}{Row variable name (optional)}

\item{v2}{Column variable name (optional)}

\item{alp}{Level of significance, Default: 0.05}

\item{prt}{Logical, whether to print tables in detail, Default: TRUE}

\item{plot}{Logical value for plotting the test result, Default: TRUE}

\item{dig}{Number of digits below the decimal point for the table, Default: 2}

\item{dig2}{Number of digits below the decimal point for the results, Default: 4}
}
\value{
list(stat=test stat, df=df, crv=critical value, pv=p-value, exp=exp table, ress=residual square)
}
\description{
Test of Homogeneity / Independence from a Cross Table
}
\examples{
x = c(39,18,12,31,14, 35,23,18,35,13,  27,16,17,24,8,  9,12,8,19,22)
x = matrix(x, nrow=4, ncol=5, byrow=TRUE)
Subject = c("Kor", "Eng", "Math", "Etc")
Hope = c("Sam", "Pub", "Exp", "Sal", "Etc")
cross.test(x)

data(Hospital, package="vcd")
cross.test(Hospital)
}
