% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{element}
\alias{element}
\title{Elements of an Event}
\usage{
element(A, r = 10)
}
\arguments{
\item{A}{Event in data frame.}

\item{r}{Maximum number of elements in a line. Default: 10}
}
\value{
Vector of elements of the event.
}
\description{
Displaying Elements of an Event
}
\examples{
S = prob::rolldie(2)
B = subset(S, (X1+X2) >=8)
element(B)

S = prob::tosscoin(4)
S=S[order(S[1], S[2], S[3], S[4]), ]
A = subset(S, apply(S, 1, function(x) sum(x=="H") >=2))
element(A, r=6)
}
