% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{freq.table}
\alias{freq.table}
\title{Create a Frequency Table}
\usage{
freq.table(x, cuts, dig = 4, mp = FALSE, ...)
}
\arguments{
\item{x}{Data vector (or matrix).}

\item{cuts}{Breaks (or number) of intervals.}

\item{dig}{Number of digits. Defaults to 4.}

\item{mp}{Logical variable for drawing a histogram. Defaults to FALSE.}

\item{...}{Other graphic parameters.}
}
\value{
xtab Frequency Table
}
\description{
To create a frequency table for a numeric vector (or matrix) x.
}
\examples{
set.seed(1234)
freq.table(rnorm(100))
freq.table(rnorm(100), mp=TRUE, prob=TRUE, col=7)
}
\keyword{Ch2}
\keyword{Descriptive}
\keyword{Statistics}
