% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{indep.event}
\alias{indep.event}
\title{Independence of Random Variables}
\usage{
indep.event(X, Y, N, ep = 1e-06)
}
\arguments{
\item{X}{First random variable.}

\item{Y}{Second random variable.}

\item{N}{Size of the sample space.}

\item{ep}{Precision limit. Default: 1e-06}
}
\value{
Probablities to be compared.
}
\description{
Determining Independence of Two Discrete Random Variables
}
\examples{
S=prob::rolldie(3); (N=nrow(S))
even=function(x) (sum(x)\%\%2)==0
span5=function(x) (max(x)-min(x))==5
A = prob::subset(S, apply(S,1,even)); nrow(A)
B = prob::subset(S, apply(S,1,span5)); nrow(B)
indep.event(A, B, nrow(S))
}
