% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{location.est}
\alias{location.est}
\title{Estimation of Location}
\usage{
location.est(x, tr = 0.1, detail = FALSE, dig = 4)
}
\arguments{
\item{x}{Vector of input data.}

\item{tr}{Treaming ratio. Default: 0.1.}

\item{detail}{Print detailed output? Default: FALSE.}

\item{dig}{Number of digits below the decimal point. Default: 4}
}
\value{
Six measures of location.
}
\description{
To calculate location estimates such as
Sample mean, median, mode, geometric mean, harmonic mean, and trimmed mean.
}
\examples{
location.est(mtcars$mpg, detail=T)

set.seed(1234)
x = round(rnorm(100, 10, 2), 1)
location.est(x, detail=T)
}
\keyword{Ch2.}
\keyword{Descriptive}
\keyword{Statistics}
