% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{mcorr}
\alias{mcorr}
\title{Multiple Scatter Plots}
\usage{
mcorr(X, item, xl, yl, mt, plot = TRUE, dig = 4)
}
\arguments{
\item{X}{List vector of bivariate data with 2 columns.}

\item{item}{String vector of list names.}

\item{xl}{Label of x-axis, Default: "group1".}

\item{yl}{Label of y-axis, Default: "group2".}

\item{mt}{Title of the scatter plot.}

\item{plot}{Logical value for drawing scatter plots, Default: TRUE.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
None.
}
\description{
Multiple Scatter Plots & Calculate Correlation Coefficients from Lists of Bivariate Data.
}
\examples{
dd = list(cbind(iris[[1]], iris[[2]]), cbind(iris[[3]], iris[[4]]))
sp = c("Sepal", "Petal")
mcorr(X=dd, item=sp, xl="Length", yl="Width")

data(exa14_1); dat=exa14_1
Subject = c("Logic", "English", "Math", "Society", "Science", "Art")
xd = list()
for (k in 1:6) xd[[k]] = cbind(dat[[2*k-1]], dat[[2*k]])
mcorr(X=xd, item=Subject, xl="1st Semester", yl="2nd Semester")
}
