% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{mcorr.test}
\alias{mcorr.test}
\title{Test for Correlations}
\usage{
mcorr.test(X, item, alp = 0.05, plot = TRUE, dig = 4)
}
\arguments{
\item{X}{List vector of bivariate data with 2 columns.}

\item{item}{String vector of list names.}

\item{alp}{Level of significance, Default: 0.05.}

\item{plot}{Logical value for drawing scatter plots, Default: TRUE.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
Table of statistics.
}
\description{
Test for Correlations Using Lists of Bivariate Data.
}
\examples{
dd = list(cbind(iris[[1]], iris[[2]]), cbind(iris[[3]], iris[[4]]))
sp = c("Sepal", "Petal")
mcorr.test(X=dd, item=sp)

data(exa14_1); dat=exa14_1
Subject = c("Logic", "English", "Math", "Society", "Science", "Art")
xd = list()
for (k in 1:6) xd[[k]] = cbind(dat[[2*k-1]], dat[[2*k]])
mcorr.test(X=xd, item=Subject)
}
