% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{meanpower.plot}
\alias{meanpower.plot}
\title{Power Function of the Test for a Mean}
\usage{
meanpower.plot(mu0, mu1, sig, nv, alp = 0.05, prng, side = "two", mt,
  dig = 4)
}
\arguments{
\item{mu0}{Population mean value under the null hypothesis, Default: 0}

\item{mu1}{Alternative population mean values, Default: NULL}

\item{sig}{Population standard deviation, Default: 1}

\item{nv}{Sample size vector, Default: c(10,30,50,100)}

\item{alp}{Level of significance, Default: 0.05}

\item{prng}{Range of x-axis}

\item{side}{Type of the alternative hypothesis ("up", "low", "two"), Default: 'two'}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Power Function of the Test for a Population Mean
}
\examples{
meanpower.plot()
meanpower.plot(mu1=0.3)

meanpower.plot(mu0=100, mu1=102, sig=5, side="up")
n = c(10, 30, 50, 70, 100)
meanpower.plot(mu0=100, mu1=102, sig=5, nv=n)

meanpower.plot(nv=50:150)
}
