% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{mreg.lse}
\alias{mreg.lse}
\title{Calculate Least Square Estimates: Multiple Regression}
\usage{
mreg.lse(form, plot = TRUE, detail = T, dig = 4)
}
\arguments{
\item{form}{Formula of regression model (ex: y ~ x1 + x2)}

\item{plot}{Logical value for drawing a scatter plot matrix, Default: TRUE.}

\item{detail}{Logical value for printing detailed calculations, Default: TRUE.}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
Matrices for calculating the LSE.
}
\description{
Calculate Least Square Estimates for a Multiple Regression.
}
\examples{
attach(mtcars)
form = mpg ~ hp + drat + wt
mreg.lse(form)

data(exa14_10)
Score=exa14_10[[1]]; Study=exa14_10[[2]]; Reading=exa14_10[[3]]
form2 = Score ~ Study + Reading
mreg.lse(form2)
}
