% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{mreg.pred}
\alias{mreg.pred}
\title{Prediction and Confidence Interval: Multiple Regression}
\usage{
mreg.pred(form, newd, pvx = 1, xrng, nx = 50, plot = F, alp = 0.05,
  dig = 4)
}
\arguments{
\item{form}{Formula of regression model (ex: y ~ x1 + x2)}

\item{newd}{A new data frame of independent variables.}

\item{pvx}{Designated number of independent variables, Default: 1.}

\item{xrng}{Plot range of the independent variable.}

\item{nx}{Number of plotting points, Default: 50.}

\item{plot}{Logical value for drawing conf. & pred. bands, Default: FALSE.}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Prediction and Confidence Interval in a Multiple Regression Model.
}
\examples{
attach(mtcars)
form = mpg ~ hp + drat + wt
newd = data.frame(hp=300, drat=4, wt=4)
mreg.pred(form, newd)
mreg.pred(form, plot=T, pvx=1, xrng=c(30,400))
mreg.pred(form, plot=T, pvx=2, xrng=c(2,6))
mreg.pred(form, plot=T, pvx=3, xrng=c(1,6))

aa = mreg.aov(form)
mreg.pred(aa, newd)
mreg.pred(aa, plot=T, pvx=1, xrng=c(30,400))

mreg.pred(lm(mpg ~ hp + drat + wt), newd)
mreg.pred(lm(mpg ~ hp + drat + wt), plot=T, pvx=1, xrng=c(30,400))

data(exa14_10)
Score=exa14_10[[1]]; Study=exa14_10[[2]]; Reading=exa14_10[[3]]
form2 = Score  Study + Reading
nd= data.frame(Study=5, Reading=5)
mreg.pred(form2, newd=nd)
mreg.pred(form2, xrng=c(0, 15), pvx=1, plot=T)
mreg.pred(form2, xrng=c(0, 15), pvx=2, plot=T)

bb = mreg.aov(form2)
mreg.pred(bb, nd)
mreg.pred(bb, xrng=c(0, 15), pvx=1, plot=T)

mreg.pred(lm(form2), nd)
mreg.pred(lm(form2), xrng=c(0, 15), pvx=1, plot=T)
}
