% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{mult.hist}
\alias{mult.hist}
\title{Multiple Histograms}
\usage{
mult.hist(df, br, mt, xl, vcol, ...)
}
\arguments{
\item{df}{Data frame of input data.}

\item{br}{Break method Default: Sturges.}

\item{mt}{Vector of main titles.}

\item{xl}{Vector of x labels.}

\item{vcol}{Vector of colors for the histograms.}

\item{...}{Other graphic parameters.}
}
\value{
None.
}
\description{
To draw multiple histograms from a data frame.
}
\examples{
mult.hist(iris[1:4])
mult.hist(iris[1:4], vcol=rainbow(4))
mult.hist(mtcars[1:7], vcol=rainbow(7))
}
\keyword{Ch2.}
\keyword{Descriptive}
\keyword{Statistics}
