% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{reg1.aov}
\alias{reg1.aov}
\title{Test for the Significance of a Simple Regression Model}
\usage{
reg1.aov(x, y, alp = 0.05, dig = 4)
}
\arguments{
\item{x}{Vector of independent variable (explanatory variable) data.}

\item{y}{Vector of dependent variable (response variable) data.}

\item{alp}{Level of significance, Default: 0.05.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
ANOVA table.
}
\description{
Test for the Significance of a Simple Regression Model by ANOVA.
}
\examples{
reg1.aov(iris[[1]], iris[[2]])
reg1.aov(iris[[3]], iris[[4]])
}
