% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{reg1.cor}
\alias{reg1.cor}
\title{Test for Correlation}
\usage{
reg1.cor(x, y, r0 = 0, alp = 0.05, plot = TRUE, dig = 4)
}
\arguments{
\item{x}{Vector of independent variable (explanatory variable) data.}

\item{y}{Vector of dependent variable (response variable) data.}

\item{r0}{Correlation coefficient value under the null hypothesis, Default: 0.}

\item{alp}{Level of significance, Default: 0.05.}

\item{plot}{Logical value for drawing scatter plots, Default: TRUE.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
None.
}
\description{
Test for Correlation in a Simple Linear Regression.
}
\examples{
reg1.cor(iris[[1]], iris[[2]])
reg1.cor(iris[[3]], iris[[4]])

x = c(1095, 1110, 1086, 1074, 1098, 1105, 1163, 1124, 1088, 1064)
y = c(49, 52, 48, 49, 50, 51, 50, 51, 49, 48)
reg1.cor(x, y)
(y2 = y*x/1000)
reg1.cor(x, y2)
reg1.cor(x, y2, r0=0.9, plot=F)
}
