% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{reg1.diag}
\alias{reg1.diag}
\title{Draw Diagnosis Plots}
\usage{
reg1.diag(x, y, wh = c(1:3, 5))
}
\arguments{
\item{x}{Vector of independent variable (explanatory variable) data.}

\item{y}{Vector of dependent variable (response variable) data.}

\item{wh}{Index vector for the diagnosis plots, Default: c(1:3,5).}
}
\value{
None.
}
\description{
Draw Diagnosis Plots of a Simple Regression Mode.
}
\examples{
reg1.diag(iris[3], iris[4])
reg1.diag(iris[3], iris[4], wh=1:2)

x = c(1095, 1110, 1086, 1074, 1098, 1105, 1163, 1124, 1088, 1064)
y = c(49, 52, 48, 49, 50, 51, 50, 51, 49, 48)
(y2 = y*x/1000)
reg1.diag(x, y2)

rr = reg1.aov(x, y2)
reg1.diag(rr)

reg1.diag(lm(y2 ~ x))
}
