% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{reg1.lse}
\alias{reg1.lse}
\title{Calculate the Least Square Estimates}
\usage{
reg1.lse(x, y, dig = 4)
}
\arguments{
\item{x}{Vector of independent variable (explanatory variable) data.}

\item{y}{Vector of dependent variable (response variable) data.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
None.
}
\description{
Calculate the Least Square Estimates for a Simple Regression.
}
\examples{
reg1.lse(iris[1], iris[3])
reg1.lse(iris[3], iris[4])

x = c(1095, 1110, 1086, 1074, 1098, 1105, 1163, 1124, 1088, 1064)
y = c(49, 52, 48, 49, 50, 51, 50, 51, 49, 48)
(y2 = y*x/1000)
reg1.lse(x, y2)
}
