% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{reg1.plot}
\alias{reg1.plot}
\title{Scatter Plot with a Simple Linear Regression Line}
\usage{
reg1.plot(x, y, xl, yl, mt, pres = TRUE, dig = 4)
}
\arguments{
\item{x}{Vector of independent variable (explanatory variable) data.}

\item{y}{Vector of dependent variable (response variable) data.}

\item{xl}{Label of x-axis (default=x variable name).}

\item{yl}{Label of y-axis (default=y variable name).}

\item{mt}{Title of the scatter plot.}

\item{pres}{Logical value for plotting residual symbols, Default: TRUE.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
None.
}
\description{
Draw a Scatter Plot with a Simple Linear Regression Line.
}
\examples{
reg1.plot(iris[[3]], iris[[4]])
reg1.plot(iris[[3]], iris[[4]], pres=F)

x = c(1095, 1110, 1086, 1074, 1098, 1105, 1163, 1124, 1088, 1064)
y = c(49, 52, 48, 49, 50, 51, 50, 51, 49, 48)
(y2 = y*x/1000)
reg1.plot(x, y2, xl="Exchange Rate", yl="Exports")
}
