% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{reg1.pre}
\alias{reg1.pre}
\title{Preliminary Analysis for a Simple Linear Regression}
\usage{
reg1.pre(x, y, xl, yl, mt, plot = TRUE, dig = 4)
}
\arguments{
\item{x}{Vector of independent variable (explanatory variable) data.}

\item{y}{Vector of dependent variable (response variable) data.}

\item{xl}{Label of x-axis (default=x variable name).}

\item{yl}{Label of y-axis (default=y variable name).}

\item{mt}{Title of the scatter plot.}

\item{plot}{Logical value for drawing scatter plots, Default: TRUE.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
None.
}
\description{
Preliminary Analysis via a Scatter Plot for a Simple Linear Regression.
}
\examples{
reg1.pre(iris[[1]], iris[[2]])
reg1.pre(iris[[3]], iris[[4]])

x = c(1095, 1110, 1086, 1074, 1098, 1105, 1163, 1124, 1088, 1064)
y = c(49, 52, 48, 49, 50, 51, 50, 51, 49, 48)
reg1.pre(x, y)
(y2 = y*x/1000)
reg1.pre(x, y2)
}
