% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{reg1.pred}
\alias{reg1.pred}
\title{Prediction and Confidence Interval}
\usage{
reg1.pred(x, y, xl, yl, mt, plot = FALSE, x0, xrng, by, alp = 0.05,
  dig = 4)
}
\arguments{
\item{x}{Vector of independent variable (explanatory variable) data.}

\item{y}{Vector of dependent variable (response variable) data.}

\item{xl}{Label of x-axis (default=x variable name).}

\item{yl}{Label of y-axis (default=y variable name).}

\item{mt}{Plot title.}

\item{plot}{Logical value for confidence & predition bands, Default: FALSE.}

\item{x0}{Conditioning value of the independent variable.}

\item{xrng}{Plot range of the independent variable.}

\item{by}{Plot interval of the independent variable.}

\item{alp}{Level of significance, Default: 0.05.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
None.
}
\description{
Prediction and Confidence Interval in a Simple Regression.
}
\examples{
reg1.pred(iris[3], iris[4], x0=2)
reg1.pred(iris[3], iris[4], x0=2, plot=T)

x = c(1095, 1110, 1086, 1074, 1098, 1105, 1163, 1124, 1088, 1064)
y = c(49, 52, 48, 49, 50, 51, 50, 51, 49, 48)
(y2 = y*x/1000)
reg1.pred(x, y2, x0=1200, plot=T)

rr = reg1.aov(x, y2)
reg1.pred(rr, x0=1200, plot=T)

reg1.pred(lm(y2 ~ x), x0=1200, plot=T)
}
