% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{reg1.test}
\alias{reg1.test}
\title{Test on the Simple Regression Coefficients}
\usage{
reg1.test(x, y, alp = 0.05, dig = 4)
}
\arguments{
\item{x}{Vector of independent variable (explanatory variable) data.}

\item{y}{Vector of dependent variable (response variable) data.}

\item{alp}{Level of significance, Default: 0.05.}

\item{dig}{Number of digits below the decimal point, Default: 4.}
}
\value{
None.
}
\description{
Test for the Significance of Simple Regression Coefficients.
}
\examples{
reg1.test(iris[1], iris[2])
reg1.test(iris[3], iris[4])

x = c(1095, 1110, 1086, 1074, 1098, 1105, 1163, 1124, 1088, 1064)
y = c(49, 52, 48, 49, 50, 51, 50, 51, 49, 48)
(y2 = y*x/1000)
reg1.test(x, y2)

rr = reg1.aov(x, y2)
reg1.test(rr)

reg1.test(lm(y2 ~ x))
}
