% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{scat.lm}
\alias{scat.lm}
\title{Scatter Plot with a Regression Line}
\usage{
scat.lm(x, y, mt, xl, yl, w = c(7, 5), ...)
}
\arguments{
\item{x}{Vector of x variable.}

\item{y}{Vector of y variable.}

\item{mt}{Title of the plot.}

\item{xl}{Label of x-axis.}

\item{yl}{Label of y-axis.}

\item{w}{Size of graphic window. Default: c(7, 5).}

\item{...}{Other graphic parameters.}
}
\value{
Object lm(y~x)
}
\description{
To create a scatter plot with a linear regression line.
}
\examples{
with(mtcars, scat.lm(wt, mpg, mt="Weight vs. MPG"))

yl = "Miles per Gallon (mpg)"
mt = paste("Scatter Plot of", names(mtcars)[3:7], "vs. MPG")
for (k in 3:7) scat.lm(mtcars[[k]],mtcars$mpg, mt=mt[k-2], xl=names(mtcars)[k], yl=yl)
}
\keyword{Ch2.}
\keyword{Descriptive}
\keyword{Statistics}
