% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{snorm.quant}
\alias{snorm.quant}
\alias{snorm.quant0}
\alias{snorm.quant2}
\title{Quantile Plot of the Standard Normal Distribution}
\usage{
snorm.quant(pv, pv2, mt, dig = 4)

snorm.quant0(pv, pv2, mt, dig = 4)

snorm.quant2(pv, pv2, mt, dig = 4)
}
\arguments{
\item{pv}{Vector of probability values}

\item{pv2}{Vector of specific probability values}

\item{mt}{Graph title (default="Quantiles of the Standard Normal Distribution")}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{pv}{Vector of probability values}

\item{pv2}{Vector of specific probability values}

\item{mt}{Graph title (default="Quantiles of the Standard Normal Distribution")}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{pv}{Vector of probability values}

\item{pv2}{Vector of specific probability values}

\item{mt}{Graph title (default="Quantiles of the Standard Normal Distribution")}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.

None.

None.
}
\description{
Quantile Plot of the Standard Normal Distribution

Quantile Plot of the Standard Normal Distribution (CDF version)

Quantile Plot of the Standard Normal Distribution (revised)
}
\examples{
pv = c(0.005, 0.01, 0.025, 0.05, 1:9/10, 0.95, 0.975, 0.99, 0.995)
snorm.quant(pv, pv)

# CDF type
snorm.quant0(pv, pv)

# Step type
snorm.quant2(pv, pv)
pv = c(0.005, 0.01, 0.025, 0.05, 1:9/10, 0.95, 0.975, 0.99, 0.995)
snorm.quant0(pv, pv)
pv = c(0.005, 0.01, 0.025, 0.05, 1:9/10, 0.95, 0.975, 0.99, 0.995)
pv2 = c(0.005, 0.025, 0.05, 0.95, 0.975, 0.995)
snorm.quant2(pv, pv2)
}
