% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{spread.est}
\alias{spread.est}
\title{Spread Estimation}
\usage{
spread.est(x, detail = FALSE, dig = 4)
}
\arguments{
\item{x}{Vector of input data.}

\item{detail}{Print detailed output? Default: FALSE.}

\item{dig}{Number of digits below the decimal point. Default: 4}
}
\value{
Five measures of dispersion.
}
\description{
To calculate spread estimates.
}
\examples{
spread.est(mtcars$mpg, detail=T)

set.seed(1234)
x = round(rnorm(100, 10, 2), 1)
spread.est(x, detail=T)
}
\keyword{Ch2.}
\keyword{Descriptive}
\keyword{Statistics}
