% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{strat.hist}
\alias{strat.hist}
\title{Stratified Histograms}
\usage{
strat.hist(ng = 3, n = 200, m = 2 + 4 * 1:ng, s = 1, sp = c(min(m) -
  2, max(m) + 2), vc, prob = FALSE)
}
\arguments{
\item{ng}{Number of classes. Defaults to 3.}

\item{n}{Number of samples. Defaults to 200.}

\item{m}{Vector of means. Defaults to 2+4*1:ng.}

\item{s}{Standard deviations. Default: 1.}

\item{sp}{Specification limits. Default: c(min(m)-2, max(m)+2).}

\item{vc}{Colors of histograms.}

\item{prob}{Density scale? Default: FALSE.}
}
\value{
None
}
\description{
To create Stratified Histograms from a mixed data set.
}
\examples{
strat.hist()
strat.hist(ng=4, m=c(6,8,12,13), sp=c(4,15))
strat.hist(ng=5, n=400, s=c(4,2,0,3,5)/10+1, prob=T)
}
\keyword{Ch2}
\keyword{Descriptive}
\keyword{Statistics.}
